cli: extract stress test result reporting into results module
Rollback trigger: issues
